File for creating .nc files with spatial correlations
This file is only to show how the spatial correlation values have been calculated. In this file, these values, as well as correlation lengths, RMSE for the exponential fits and distances are saved as .nc files, which are used in the  "Spatio-temporal correlations - only plots" notebook. The resulting .nc files are saved in this repository in the "Spatial Correlation for corrected CMIP6 models" folder.